refactor: split processor and escrow tests
